Add ens contracts as submodule, cleanup, update README
Name bazaar is dependent on ENS contracts (ensdomains/ens repository). This PR adds the the repo as a submodule
fixed at a specific version instead of copied contracts from ens. This should improve contract migrations in the
future, because one can just bump the version of submodule.

We have also reflected these changes in README and updated gitignore to avoid commiting generated files.

Unfortunately, my IDE made a few formatting changes as well, which are included in this PR as well.